backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
python-dotenv==1.0.1
django-cors-headers==4.4.0
requests==2.32.3
gunicorn==22.0.0
reportlab==4.2.2
psycopg2-binary==2.9.9
